[1/3] AST serialization, added new serialization

+++ Apps/Editor/Include/Tools/BigBestFitArenaDebugger.h
@@ -34,7 +34,7 @@ namespace Rift
 
 		u32 GetY(sizet offset)
 		{
-			return offset / bytesPerRow;
+			return u32(offset / bytesPerRow);
 		}
 
 		u32 GetX(sizet offset)
